move to to util
